+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## 1.1 branch
 ### 1.1.7
 * added `is_true_or_fail()` function;
+* added `create_file()` and `create_tmp_file()` global functions;
 * `file_exists_or_fail()`, `is_dir_or_fail()`, `is_readable_or_fail()` and
     `is_writable_or_fail()` functions are now deprecated and they will be
     removed from the `1.2.0` version. Use instead `is_true_or_fail()`.